Add settings for ghost colors
Both entities and tiles have colors, and different colors for ghosts
that have a construction bot scheduled. There is also a color for wire
ghosts.

The game also differentiates between "shaders" and "shaderless" ghosts.
The documentation does not explain what the difference is at all. Rather
than confuse users, just set both with the same colors, even though the
default colors are different.

That was added in 2.0.15, and before then, wire ghost color did not
exist either. This adds a lot of ugly code to handle old versions.

Also, until 2.0.55, there was no way to check the game version (base
version yes, but it is possible to disable the base mod), so we just
have to show the wire ghost color setting always even when it will do
nothing.

Resolves: #14